refactor: Extract AddMenuContent from AddMenu
no code change
